requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
idna==2.10
pymongo==3.11.1
pytz==2020.4
requests==2.25.0
sqlparse==0.2.4
urllib3==1.26.2
